requirements.txt
librosa==0.8
pytorch=1.15
numpy
fastapi
matplotlib